use floating pt math for percentages

+++ src/java/org/apache/cassandra/db/ColumnFamilyStore.java
@@ -1919,7 +1919,7 @@ public class ColumnFamilyStore implements ColumnFamilyStoreMBean
     public float getPercentageTombstonesPerLastRead()
     {
         long total = metric.tombstoneScannedHistogram.count() + metric.liveScannedHistogram.count();
-        return (metric.tombstoneScannedHistogram.count() / total);
+        return ((float) metric.tombstoneScannedHistogram.count() / total);
     }
 
     public long getLiveCellsPerLastRead()
